// src/shared/ui/Typography/typographyVariant.enum.ts
export enum TypographyVariant {
  Large = 'large',
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  RegularText16 = 'regular-text-16',
  BoldText16 = 'bold-text-16',
  RegularText14 = 'regular-text-14',
  MediumText14 = 'medium-text-14',
  BoldText14 = 'bold-text-14',
  SmallText = 'small-text',
  SemiBoldSmallText = 'semi-bold-small-text',
  RegularLink = 'regular-link',
  SmallLink = 'small-link',
}